fix(core): normalize multimodal file handling config
Move FileHandlingConfig to platform types and consume it from middleware to avoid cross-module type drift. Also apply per-message max size overrides from file input attrs and remove stale inline-data branching.

Refs: none
